place order Api created

+++ Backend/Electronic-shopping-backend/src/main/java/com/backend/user/Repository/UserRepository.java
@@ -15,6 +15,6 @@ public interface UserRepository extends JpaRepository<User , Long>{
     Optional<User> findByEmail(String email);
     
     
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 
 }
