+++ Backend/Electronic-shopping-backend/src/main/java/com/backend/user/service/UserServiceImpl.java
@@ -77,7 +77,7 @@ public class UserServiceImpl implements UserService {
     
     
     @Override
-    public UserDetailsResponseDTO getUserById(Integer id) {
+    public UserDetailsResponseDTO getUserById(Long id) {
 
         User user = userRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
@@ -93,7 +93,7 @@ public class UserServiceImpl implements UserService {
     }
     
     @Override
-    public UserDetailsResponseDTO updateUser(Integer id, UpdateUserDTO dto) {
+    public UserDetailsResponseDTO updateUser(Long id, UpdateUserDTO dto) {
 
         User user = userRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
